refactor: tinyId generator
